+++ manul/src/combinators.rs
@@ -1,0 +1,3 @@
+//! Combinators operating on protocols.
+
+pub mod misbehave;