+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PartyId` trait alias for the combination of bounds needed for a party identifier. ([#59])
 - An impl of `ProtocolError` for `()` for protocols that don't use errors. ([#60])
 - A dummy `CorrectnessProof` trait. ([#60])
+- A `misbehave` combinator, intended primarily for testing. ([#60])
 
 
 [#32]: https://github.com/entropyxyz/manul/pull/32